+++ engine/render/source/render/particles/generators/frame_particle_system_generator.cpp
@@ -9,8 +9,8 @@ namespace kw {
 
 ParticleSystemGenerator* FrameParticleSystemGenerator::create_from_markdown(MemoryResource& memory_resource, ObjectNode& node) {
     return memory_resource.construct<FrameParticleSystemGenerator>(
-        static_cast<float>(node["min"].as<NumberNode>().get_value()),
-        static_cast<float>(node["max"].as<NumberNode>().get_value())
+        node["min"].as<NumberNode>().get_value(),
+        node["max"].as<NumberNode>().get_value()
     );
 }
